+++ code/source/draw.h
@@ -1,10 +1,10 @@
 #ifndef PONG_DRAW_H
 #define PONG_DRAW_H
 
-void drawline(u8 x0, u8 y0, u8 x1, u8 y1, int color);
-void drawRectXYXY(u8 x0, u8 y0, u8 x1, u8 y1, int color);
-void drawRectXYHW(u8 x, u8 y, u8 h, u8 w, int color);
-void drawRectXYHWfill(u8 x, u8 y, u8 h, u8 w, int color);
-void drawCubeCentered(u8 x, u8 y, u8 h, int color);
+void drawline(int x0, int y0, int x1, int y1, int color);
+void drawRectXYXY(int x0, int y0, int x1, int y1, int color);
+void drawRectXYHW(int x, int y, int h, int w, int color);
+void drawRectXYHWfill(int x, int y, int h, int w, int color);
+void drawCubeCentered(int x, int y, int h, int color);
 
 #endif
